各ファイルに書いていたテストをtestディレクトリに移行
