+++ src/exec_cmd/exec_cmd.c
@@ -18,31 +18,3 @@ void	exec_cmd(t_process **procs)
 		++i;
 	}
 }
-
-#ifdef EXEC_CMD_C
-
-#include "main.h"
-#include "parse.h"
-#include "utils.h"
-#include "debug.h"
-#include "constants.h"
-#include "read_cmd_line.h"
-#include "libft.h"
-
-int main(void)
-{
-	char cmd_line[ARG_MAX + 1];
-	t_process **cmd_procs;
-
-	while(TRUE)
-	{
-		print_prompt();
-		read_cmd_line(cmd_line);
-		cmd_procs = parse_cmd_line(cmd_line, &g_status);
-		exec_cmd(cmd_procs);
-		ft_putnbr_fd(WEXITSTATUS(g_status), 1);
-		pendl();
-	}
-}
-
-#endif
