+++ src/exec_cmd/exec_pipes.c
@@ -108,36 +108,3 @@ void		exec_pipes(t_process *procs)
 	wait_child_procs(pid, count_procs(procs));
 	free_pipe_fd_array(pipe_fd, procs);
 }
-
-#ifdef EXEC_PIPES_C
-
-#include "main.h"
-#include "parse.h"
-#include "debug.h"
-#include "utils.h"
-#include "libft.h"
-#include "env_ctrl.h"
-#include "read_cmd_line.h"
-
-int main(int argc, char *argv[], char *envp[])
-{
-	// int status;
-	// int pid;
-	char cmd_line[ARG_MAX + 1];
-	t_process **cmd_procs;
-
-	argc += 1;
-	argv[0][0] = 'a';
-	create_env_list(envp);
-	while(TRUE)
-	{
-		print_prompt();
-		read_cmd_line(cmd_line);
-		cmd_procs = parse_cmd_line(cmd_line, &g_status);
-		exec_pipes(cmd_procs[0]);
-		ft_putnbr_fd(WEXITSTATUS(g_status), 1);
-		pendl();
-	}
-}
-
-#endif
