+++ test/test_env.c
@@ -1,3 +1,4 @@
+#include "env_ctrl.h"
 #include "struct/env_list.h"
 
 #ifdef CREATE_ENV_LIST_C
@@ -15,6 +16,26 @@ int	main(int argc, char *argv[], char *envp[])
 		printf("%s=%s\n", env_node->key, env_node->value);
 		env_node = env_node->next;
 	}
+}
+
+#endif
+
+
+#ifdef 	MAKE_NEW_ENV_NODE_C
+
+#include <stdio.h>
+
+int main(void)
+{
+	t_env_list *new;
+	char test_env[] = "foo=foo";
+
+	new = make_new_env_node(test_env);
+	printf("key:%s\n", new->key);
+	printf("value:%s\n", new->value);
+	free(new->key);
+	free(new->value);
+	free(new);
 }
 
 #endif
